requirements_cpu.txt
bleach==1.5.0
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
imageio==2.2.0
Markdown==2.6.11
numpy==1.14.2
Pillow==5.0.0
protobuf==3.5.2
six==1.11.0
tensorflow-cpu==1.4.1
tensorflow-tensorboard==0.4.0
tqdm==4.19.7
Werkzeug==0.14.1
